Rick visual stuff
* messing around with the stored procedures

pretty sure I'm a retard because I can't get the ending stack field to
take from a variable, but it works fine from a value

* Stored Procedure fixed

* Revert "Stored Procedure fixed"

This reverts commit 16305e6a8f81b8c7a949459ce6a74a3149fe1069.

* Revert "Revert "Stored Procedure fixed""

This reverts commit b2993bf5a357ac1d8262034bc5adecbcf1071895.

* Theming changes

* More layout stuff

* Lots of Layout and theming changes

Open to ideas on this stuff, but this is definitely better than the
original

* Some more minor layout changes

putting buttons to the theme

* More visual changes

tribar moved to left. buttons gradient removed. transform added to
buttons

* Revert "More visual changes"

This reverts commit 27b464735d2b111f650bc47342a896d291345b6a.

* more visual changes